# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Software Engineering School 4.0 Test API",
        "version": "0.1"
    },
    "paths": {
        "/api/rate": {
            "get": {
                "tags": [
                    "Currency"
                ],
                "summary": "Get current currency rate",
                "description": "Get current currency rate",
                "operationId": "get-current-rate",
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CurrentRateResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid status value"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/subscribe": {
            "post": {
                "tags": [
                    "Subscription"
                ],
                "summary": "Subscribe the email",
                "description": "Subscribe the email",
                "operationId": "subscribe-the-email",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SubscribeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Email successfully added"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Email already exists"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "SubscribeRequest": {
                "title": "Subscribe request",
                "description": "Subscribe request body data",
                "required": [
                    "email"
                ],
                "properties": {
                    "email": {
                        "description": "User email",
                        "type": "string",
                        "example": "test@test.com"
                    }
                },
                "type": "object"
            },
            "CurrentRateResource": {
                "title": "CurrentRateResource",
                "properties": {
                    "buy": {
                        "description": "Current buy currency rate",
                        "type": "number",
                        "format": "float",
                        "example": 39.5
                    },
                    "sale": {
                        "description": "Current sale currency rate",
                        "type": "number",
                        "format": "float",
                        "example": 41.5
                    }
                },
                "type": "object"
            }
        }
    }
}